remove unused screens and move asset loading to preloader for city
