--- src/game/scenes/GameOver.js
@@ -1,30 +1,0 @@
-import { EventBus } from '../EventBus';
-import { Scene } from 'phaser';
-
-export class GameOver extends Scene
-{
-    constructor ()
-    {
-        super('GameOver');
-    }
-
-    create ()
-    {
-        this.cameras.main.setBackgroundColor(0xff0000);
-
-        this.add.image(512, 384, 'background').setAlpha(0.5);
-
-        this.add.text(512, 384, 'Game Over', {
-            fontFamily: 'Arial Black', fontSize: 64, color: '#ffffff',
-            stroke: '#000000', strokeThickness: 8,
-            align: 'center'
-        }).setOrigin(0.5).setDepth(100);
-
-        EventBus.emit('current-scene-ready', this);
-    }
-
-    changeScene ()
-    {
-        this.scene.start('MainMenu');
-    }
-}
