+++ src/game/scenes/MainMenu.js
@@ -6,21 +6,16 @@ export class MainMenu extends Scene {
     }
 
     create() {
-        // Add main title
+        // TITLE
         this.add
-            .text(
-                this.cameras.main.centerX,
-                this.cameras.main.centerX - 200,
-                "Your Game Title",
-                {
-                    fontSize: "48px",
-                    color: "#ffffff",
-                    fontFamily: "Arial Black",
-                }
-            )
+            .text(this.cameras.main.centerX, 300, "RareQuest", {
+                fontSize: "72px",
+                color: "#ffffff",
+                fontFamily: "Arial Black",
+            })
             .setOrigin(0.5);
 
-        // Add play button
+        // PLAY BUTTON
         const playButton = this.add
             .text(
                 this.cameras.main.centerX,
@@ -32,13 +27,11 @@ export class MainMenu extends Scene {
                     fontFamily: "Arial Black",
                 }
             )
-            .setOrigin(0.5);
+            .setOrigin(0.5)
+            .setInteractive();
 
-        // Define action for when the play button is clicked
-        playButton.setInteractive();
         playButton.on("pointerdown", () => {
-            // Start the "ChooseCharacter" scene when the play button is clicked
-            this.scene.start("ChooseCharacter");
+            this.scene.start("City");
         });
     }
 }
