+++ src/game/scenes/index.js
@@ -1,17 +1,9 @@
 import { Boot } from "./Boot";
-import { Game } from "./Game";
-import { GameOver } from "./GameOver";
+import { City } from "./City";
+
 import { MainMenu } from "./MainMenu";
 import { Preloader } from "./Preloader";
-import { ChooseCharacter } from "./ChooseCharacter";
 
 // IMPORTANT: you must import and add your new scene to this array for it to be renderable!!
-export const scenes = [
-    Boot,
-    Preloader,
-    MainMenu,
-    ChooseCharacter,
-    Game,
-    GameOver,
-];
+export const scenes = [Boot, Preloader, MainMenu, City];
